+++ ContentPatcher/release-notes.md
@@ -1,8 +1,10 @@
 [← back to readme](README.md)
 
 # Release notes
-## Upcoming release
-* `ConfigSchema` options can now have an optional `Description` field, which will be shown in UIs like Generic Mod Config Menu.
+## 1.18.2
+Released 13 September 2020.
+
+* `ConfigSchema` options can now have an optional `Description` field, which is shown in UIs like Generic Mod Config Menu.
 * Fixed `TextOperations` incorrectly adding delimiters when there's no previous value.
 * Fixed errors sometimes showing "ContentPatcher.Framework.Conditions.TokenString" instead of the intended value.